Fix: MainPage rename for Travis build
